requirements.txt
numpy
pandas
flask
fastapi
uvicorn
scikit-learn
transformers
torch
pathlib
llama-cpp-python
pydantic-settings
chromadb
sentence-transformers
